Из aiohttp в httpx и обратно Перебрал структуру модулей, подчистил коментарии и импорты
