+++ utils/utils.py
@@ -1,5 +1,8 @@
 from datetime import datetime
 
+from keyboards.user_keyboards import system_buttons
+from file_managment.ya_file_manager import path_list
+
 subsequence = 1  # Последнее значение в нейминге файлов для уникальности имени
 
 
@@ -15,5 +18,14 @@ def get_dt_name() -> str:
     return dt_name
 
 
-def set_correct_path(path) -> str:
-    return '/'.join(path) + '/'
+def move_to_folders_on_disk(folder) -> None:
+    '''Функция осуществляет смену директорий путем добавления или удаления выбранной папки в список,
+     из которого будет сформирован корректный путь'''
+    if len(path_list) != 0 and folder in system_buttons.values():
+        if folder == system_buttons['back']:
+            path_list.pop(-1)
+        elif folder == system_buttons['home']:
+            path_list.clear()
+    if folder not in ['', *system_buttons.values()]:
+        path_list.append(folder)
+
